# ai_tnhn/ai-api-tnhn/docs/swagger.yaml
basePath: /api
definitions:
...
  models.User:
    properties:
      active:
        example: true
        type: boolean
      assigned_emergency_construction_ids:
        items:
          type: string
        type: array
      assigned_inundation_station_ids:
        description: RBAC for Employee
        items:
          type: string
        type: array
      assigned_lake_station_ids:
        items:
          type: string
        type: array
      assigned_pumping_station_id:
        type: string
      assigned_rain_station_ids:
        items:
          type: string
        type: array
      assigned_river_station_ids:
        items:
          type: string
        type: array
      created_at:
        type: integer
      deleted_at:
        type: integer
      email:
        example: admin@hsdc.com.vn
        type: string
      id:
        type: string
      is_company:
        example: true
        type: boolean
      is_employee:
        example: false
        type: boolean
      name:
        example: Nguyễn Văn A
        type: string
      org_id:
        example: 60f123456789
        type: string
      password:
        example: '********'
        type: string
      role:
        example: admin
        type: string
      updated_at:
        type: integer
      username:
        example: admin
        type: string
    type: object
  web.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      error:
        example: error message
        type: string
      status:
        example: error
        type: string
    type: object
  web.Response:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      status:
        example: success
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Máy chủ API cho Hệ thống Giám sát Ngập lụt Hà Nội.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: API Hệ thống Giám sát Ngập lụt Hà Nội
  version: "1.0"
paths:
  /admin/employees:
    get:
      description: Truy xuất danh sách nhân viên với tính năng lọc
      parameters:
      - description: Lọc theo ID đơn vị
        in: query
        name: org_id
        type: string
      - description: Lọc theo tên
        in: query
        name: name
        type: string
      - description: Số trang
        in: query
        name: page
        type: integer
      - description: Số bản ghi mỗi trang
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  properties:
                    data:
                      items:
                        $ref: '#/definitions/models.User'
                      type: array
                    total:
                      type: integer
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: Danh sách nhân viên
      tags:
      - Nhân viên
    post:
      consumes:
      - application/json
      description: Tạo mới một người dùng/nhân viên hệ thống
      parameters:
      - description: Dữ liệu nhân viên
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Tạo mới một nhân viên
      tags:
      - Nhân viên
  /admin/employees/{id}:
    delete:
      description: Loại bỏ một người dùng/nhân viên theo ID
      parameters:
      - description: ID nhân viên
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Xóa nhân viên
      tags:
      - Nhân viên
    get:
      description: Truy xuất thông tin chi tiết của một người dùng/nhân viên cụ thể
      parameters:
      - description: ID nhân viên
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lấy thông tin nhân viên theo ID
      tags:
      - Nhân viên
    put:
      consumes:
      - application/json
      description: Cập nhật thông tin chi tiết của một người dùng/nhân viên hiện có
      parameters:
      - description: ID nhân viên
        in: path
        name: id
        required: true
        type: string
      - description: Dữ liệu nhân viên cập nhật
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cập nhật thông tin nhân viên
      tags:
      - Nhân viên
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"